refactored db request code into utilities
